Separate out logic and use in POA coinview rule
