+++ src/Stratis.Bitcoin.Features.SmartContracts/Rules/SmartContractCoinViewRuleLogic.cs
@@ -1,0 +1,231 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using NBitcoin;
+using Stratis.Bitcoin.Base.Deployments;
+using Stratis.Bitcoin.Consensus;
+using Stratis.Bitcoin.Consensus.Rules;
+using Stratis.Bitcoin.Features.Consensus;
+using Stratis.Bitcoin.Features.Consensus.Rules;
+using Stratis.Bitcoin.Features.Consensus.Rules.CommonRules;
+using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Core.Receipts;
+using Stratis.SmartContracts.Core.State;
+using Stratis.SmartContracts.Core.Util;
+
+namespace Stratis.Bitcoin.Features.SmartContracts.Rules
+{
+    /// <summary>
+    /// Abstraction for shared SC coinview rule logic.
+    /// </summary>
+    public sealed class SmartContractCoinViewRuleLogic
+    {
+        protected List<Transaction> blockTxsProcessed;
+        protected Transaction generatedTransaction;
+        protected IList<Receipt> receipts;
+        protected uint refundCounter;
+        protected IStateRepositoryRoot mutableStateRepository;
+
+        public ISmartContractCoinviewRule ContractCoinviewRule { get; set; }
+
+   
+        public async Task RunAsync(Func<RuleContext, Task> baseRunAsync, RuleContext context)
+        {
+            this.blockTxsProcessed = new List<Transaction>();
+            this.receipts = new List<Receipt>();
+            this.refundCounter = 1;
+            NBitcoin.Block block = context.ValidationContext.BlockToValidate;
+
+            // Get a IStateRepositoryRoot we can alter without affecting the injected one which is used elsewhere.
+            byte[] blockRoot = ((ISmartContractBlockHeader)context.ValidationContext.ChainedHeaderToValidate.Previous.Header).HashStateRoot.ToBytes();
+            this.mutableStateRepository = this.ContractCoinviewRule.OriginalStateRoot.GetSnapshotTo(blockRoot);
+
+            await baseRunAsync(context);
+
+            if (new uint256(this.mutableStateRepository.Root) != ((ISmartContractBlockHeader)block.Header).HashStateRoot)
+                SmartContractConsensusErrors.UnequalStateRoots.Throw();
+
+            ValidateAndStoreReceipts(((ISmartContractBlockHeader)block.Header).ReceiptRoot);
+
+            // Push to underlying database
+            this.mutableStateRepository.Commit();
+
+            // Update the globally injected state so all services receive the updates.
+            this.ContractCoinviewRule.OriginalStateRoot.SyncToRoot(this.mutableStateRepository.Root);
+        }
+
+        /// <summary>
+        /// Executes contracts as necessary and updates the coinview / UTXOset after execution.
+        /// </summary>
+        /// <inheritdoc/>
+        public void UpdateCoinView(Action<RuleContext, Transaction> baseUpdateUTXOSet, RuleContext context, Transaction transaction)
+        {
+            if (this.generatedTransaction != null)
+            {
+                ValidateGeneratedTransaction(transaction);
+                baseUpdateUTXOSet(context, transaction);
+                this.blockTxsProcessed.Add(transaction);
+                return;
+            }
+
+            // If we are here, was definitely submitted by someone
+            ValidateSubmittedTransaction(transaction);
+
+            TxOut smartContractTxOut = transaction.Outputs.FirstOrDefault(txOut => SmartContractScript.IsSmartContractExec(txOut.ScriptPubKey));
+            if (smartContractTxOut == null)
+            {
+                // Someone submitted a standard transaction - no smart contract opcodes.
+                baseUpdateUTXOSet(context, transaction);
+                this.blockTxsProcessed.Add(transaction);
+                return;
+            }
+
+            // Someone submitted a smart contract transaction.
+            ExecuteContractTransaction(context, transaction);
+
+            baseUpdateUTXOSet(context, transaction);
+            this.blockTxsProcessed.Add(transaction);
+        }
+
+        /// <summary>
+        /// Validates that any condensing transaction matches the transaction generated during execution
+        /// </summary>
+        /// <param name="transaction"></param>
+        private void ValidateGeneratedTransaction(Transaction transaction)
+        {
+            if (this.generatedTransaction.GetHash() != transaction.GetHash())
+                SmartContractConsensusErrors.UnequalCondensingTx.Throw();
+
+            this.generatedTransaction = null;
+
+            return;
+        }
+
+        /// <summary>
+        /// Validates that a submitted transacction doesn't contain illegal operations
+        /// </summary>
+        /// <param name="transaction"></param>
+        protected void ValidateSubmittedTransaction(Transaction transaction)
+        {
+            if (transaction.Inputs.Any(x => x.ScriptSig.IsSmartContractSpend()))
+                SmartContractConsensusErrors.UserOpSpend.Throw();
+
+            if (transaction.Outputs.Any(x => x.ScriptPubKey.IsSmartContractInternalCall()))
+                SmartContractConsensusErrors.UserInternalCall.Throw();
+        }
+
+        /// <summary>
+        /// Throws a consensus exception if the gas refund inside the block is different to what this node calculated during execution.
+        /// </summary>
+        public void ValidateRefunds(TxOut refund, Transaction coinbaseTransaction)
+        {
+            TxOut refundToMatch = coinbaseTransaction.Outputs[this.refundCounter];
+            if (refund.Value != refundToMatch.Value || refund.ScriptPubKey != refundToMatch.ScriptPubKey)
+            {
+                SmartContractConsensusErrors.UnequalRefundAmounts.Throw();
+            }
+
+            this.refundCounter++;
+        }
+
+        /// <summary>
+        /// Executes the smart contract part of a transaction
+        /// </summary>
+        protected void ExecuteContractTransaction(RuleContext context, Transaction transaction)
+        {
+            IContractTransactionContext txContext = GetSmartContractTransactionContext(context, transaction);
+            IContractExecutor executor = this.ContractCoinviewRule.ExecutorFactory.CreateExecutor(this.mutableStateRepository, txContext);
+
+            IContractExecutionResult result = executor.Execute(txContext);
+
+            var receipt = new Receipt(
+                new uint256(this.mutableStateRepository.Root),
+                result.GasConsumed,
+                result.Logs.ToArray(),
+                txContext.TransactionHash,
+                txContext.Sender,
+                result.To,
+                result.NewContractAddress,
+                !result.Revert,
+                result.ErrorMessage
+            )
+            {
+                BlockHash = context.ValidationContext.BlockToValidate.GetHash()
+            };
+
+            this.receipts.Add(receipt);
+
+            ValidateRefunds(result.Refund, context.ValidationContext.BlockToValidate.Transactions[0]);
+
+            if (result.InternalTransaction != null)
+                this.generatedTransaction = result.InternalTransaction;
+        }
+
+        /// <summary>
+        /// Retrieves the context object to be given to the contract executor.
+        /// </summary>
+        private IContractTransactionContext GetSmartContractTransactionContext(RuleContext context, Transaction transaction)
+        {
+            ulong blockHeight = Convert.ToUInt64(context.ValidationContext.ChainedHeaderToValidate.Height);
+
+            GetSenderResult getSenderResult = this.ContractCoinviewRule.SenderRetriever.GetSender(transaction, ((PowConsensusRuleEngine)this.Parent).UtxoSet, this.blockTxsProcessed);
+
+            if (!getSenderResult.Success)
+                throw new ConsensusErrorException(new ConsensusError("sc-consensusvalidator-executecontracttransaction-sender", getSenderResult.Error));
+
+            Script coinbaseScriptPubKey = context.ValidationContext.BlockToValidate.Transactions[0].Outputs[0].ScriptPubKey;
+
+            GetSenderResult getCoinbaseResult = this.ContractCoinviewRule.SenderRetriever.GetAddressFromScript(coinbaseScriptPubKey);
+
+            uint160 coinbaseAddress = (getCoinbaseResult.Success) ? getCoinbaseResult.Sender : uint160.Zero;
+
+            Money mempoolFee = transaction.GetFee(((UtxoRuleContext)context).UnspentOutputSet);
+
+            return new ContractTransactionContext(blockHeight, coinbaseAddress, mempoolFee, getSenderResult.Sender, transaction);
+        }
+
+        /// <summary>
+        /// Throws a consensus exception if the receipt roots don't match.
+        /// </summary>
+        private void ValidateAndStoreReceipts(uint256 receiptRoot)
+        {
+            List<uint256> leaves = this.receipts.Select(x => x.GetHash()).ToList();
+            bool mutated = false; // TODO: Do we need this?
+            uint256 expectedReceiptRoot = BlockMerkleRootRule.ComputeMerkleRoot(leaves, out mutated);
+
+            if (receiptRoot != expectedReceiptRoot)
+                SmartContractConsensusErrors.UnequalReceiptRoots.Throw();
+
+            // TODO: Could also check for equality of logsBloom?
+
+            this.ContractCoinviewRule.ReceiptRepository.Store(this.receipts);
+            this.receipts.Clear();
+        }
+
+        public bool CheckInput(Func<Transaction, int, TxOut, PrecomputedTransactionData, TxIn, DeploymentFlags, bool> baseCheckInput, Transaction tx, int inputIndexCopy, TxOut txout, PrecomputedTransactionData txData, TxIn input, DeploymentFlags flags)
+        {
+            if (txout.ScriptPubKey.IsSmartContractExec() || txout.ScriptPubKey.IsSmartContractInternalCall())
+            {
+                return input.ScriptSig.IsSmartContractSpend();
+            }
+
+            return baseCheckInput(tx, inputIndexCopy, txout, txData, input, flags);
+        }
+
+        public bool IsProtocolTransaction(Transaction transaction)
+        {
+            return transaction.IsCoinBase || transaction.IsCoinStake;
+        }
+
+        public void ResetRefundCounter()
+        {
+            this.refundCounter = 1;
+        }
+
+        public void ClearGeneratedTransaction()
+        {
+            this.generatedTransaction = null;
+        }
+    }
+}
